Use Redis pipelining to batch requests
This might help make this check a little faster since we replaced an MGET call
with a bunch of separate EXISTS calls. Using pipelining means we don't have to
wait for each individual request/response cycle.

Also, ensure we are always using symbols when comparing queue names. The tests
I added weren't passing before making this tweak. We convert queue name to a
symbol elsewhere but we weren't doing so here.